backend/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-firestore
google-cloud-tasks
google-cloud-storage
httpx